feat: implement terminal user interface (TUI) for event monitoring and control
